server/requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
requests==2.28.1
